Call backend invoke from BackendPathEvaluator
